+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- Display entities' media clock master id and name in the entity list
+- Media clock management dialog for easy setup/status access of media clock distribution
+- Support for media clock master determination
+- Support for media clock distribution handling, incl. entity clock source configuration and CRF stream connection establishment
 - Keyboard shortcut to refresh the controller (CTRL-R)
 - Button to refresh the controller (next to the Interface selection dropdown)